- make JL_JsvalToNativeVector support TypedArray. - more optimizations (RESTRICT, __assume in ASSERT, ...). - move more common vcproj stuff to vsprops files. - add JLCpuInfo(). add jsstd::CPUID property. - change hash::call-function behavior. Rename Hash::Init() into Hash::Reset().

+++ src/jsprotex/static.cpp
@@ -91,7 +91,7 @@ DEFINE_FUNCTION( PerlinNoise ) {
 
 	JL_S_ASSERT_ARG_RANGE(4,6);
 	int n;
-	double a, b, x, y, z;
+	float a, b, x, y, z;
 	JL_CHK( JL_JsvalToNative(cx, JL_ARG(1), &a) );
 	JL_CHK( JL_JsvalToNative(cx, JL_ARG(2), &b) );
 	JL_CHK( JL_JsvalToNative(cx, JL_ARG(3), &n) );
